chore: reorganize oracles
