+++ src/fuzzer/oracles/Types.ts
@@ -1,0 +1,1 @@
+export type Judgment = "fail" /* -1 */ | "unknown" /* 0 */ | "pass" /* 1 */;
